requirements.txt
Flask==0.10.1
urllib3==1.9.1
Pillow
certifi
pyvirtualdisplay
